Update to master

+++ server/src/Utils/StopwatchExtensions.cs
@@ -4,17 +4,17 @@ namespace KRPC.Utils
 {
     static class StopwatchExtensions
     {
-        static readonly double ticksToSeconds = 1d / (double)Stopwatch.Frequency;
-        static readonly double microsecondsToTicks = (double)Stopwatch.Frequency / 1000000d;
+        static readonly double ticksToSeconds = 1d / Stopwatch.Frequency;
+        static readonly double microsecondsToTicks = Stopwatch.Frequency / 1000000d;
 
         public static double ElapsedSeconds (this Stopwatch stopwatch)
         {
-            return (double)stopwatch.ElapsedTicks * ticksToSeconds;
+            return stopwatch.ElapsedTicks * ticksToSeconds;
         }
 
         public static long MicrosecondsToTicks (long microseconds)
         {
-            return (long)((double)microseconds * microsecondsToTicks);
+            return (long)(microseconds * microsecondsToTicks);
         }
     }
 }
